fix the cycling reference issue between api and xml_utils
